requirements/base.txt
matplotlib
numpy
pillow
scipy
nicr-scene-analysis-datasets>=0.8.3
